moved everything to separate vue file
