+++ src/css/quasar.variables.scss
@@ -12,11 +12,13 @@
 // to match your app's branding.
 // Tip: Use the "Theme Builder" on Quasar's documentation website.
 
-$primary   : #1976D2;
-$secondary : #26A69A;
+// src/css/quasar.variables.scss
+
+$primary   : #5e0413;
+$secondary : #300a7d;
 $accent    : #9C27B0;
 
-$dark      : #1D1D1D;
+$dark      : #1d1d1d;
 
 $positive  : #21BA45;
 $negative  : #C10015;
